The last change to plugins.c erroneously passes a size_t as a field width to asprintf.  Address the signed/unsigned warning cleanup using a cast instead.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@20931 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/support/plugins.c
@@ -618,7 +618,7 @@ krb5int_open_plugin_dirs (const char * const *dirnames,
                 }
                 
 		if (!err) {
-                    size_t len = NAMELEN (d);
+                    int len = NAMELEN (d);
 		    if (asprintf(&filepath, "%s/%*s", dirnames[i], len, d->d_name) < 0) {
 			filepath = NULL;
 			err = ENOMEM;
